Apply velocity to relative neighbour vector
Before, the velocity vector was dot-multiplied with the vector from the
center of the object to the cell whose border was sought. Now, we use
the offset of the neighbour cell towards the border cell, which is
better for non-straight shaped objects and removes the spherical shape
of the shockwave in front of cube objects.