New format conventions and draft of Level format
ManiacLab data files are now split in a header and a body part. Each
part is a complete structstream. This allows cheap reading of header
information for each file type.

In particular, level collections can become quite large and it is
important to be able to read each level separately (for which there is
a table of contents in the level collection header).